feat: check file permissions before accessing an open file
Print Kernel messages when the system is interacting with files that are
named 'confidential'.

Print the mask value. The mask value represents the type of operation
being requested on a file.

mask values:
MAY_READ (0x4): Indicates that the permission check is for reading the file.
MAY_WRITE (0x2): Indicates that the permission check is for writing to the file.
MAY_EXEC (0x1): Indicates that the permission check is for executing the file.
MAY_APPEND (0x8): Indicates that the permission check is for appending to the file.